backend/migrations/models/25_20241107081601_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "alert" ALTER COLUMN "state" DROP DEFAULT;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "alert" ALTER COLUMN "state" SET DEFAULT NULL;"""
