backend/requirements.txt
aerich==0.7.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.9.1
asn1crypto==1.5.1
astroid==3.2.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-compute==34.1.0
azure-mgmt-core==1.4.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-monitor==6.0.2
azure-mgmt-resource==23.1.1
azure-monitor-query==1.4.1
azure-storage-blob==12.21.0
billiard==4.2.0
boto3==1.34.145
botocore==1.34.145
cachetools==5.4.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
databases==0.9.0
db-dtypes==1.2.0
dictdiffer==0.9.0
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-azure-auth==4.4.0
fastapi-cli==0.0.4
filelock==3.15.4
flower==2.0.1
frozenlist==1.7.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.25.0
google-cloud-billing==1.13.5
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.65.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
idna==3.7
iniconfig==2.0.0
iso8601==1.1.0
isodate==0.6.1
isort==5.13.2
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
junit-xml-2==1.9
kombu==5.3.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
msal==1.30.0
msal-extensions==1.2.0
multidict==6.6.3
numpy==2.0.0
openai==1.95.0
orjson==3.10.3
packaging==24.1
pandas==2.2.2
passlib==1.7.4
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.2
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.5
pylint-junit==0.3.4
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.11.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
starlette==0.37.2
tomli==2.0.1
tomlkit==0.12.5
tornado==6.4.1
tortoise-orm==0.21.3
tqdm==4.67.1
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.20.1
